Synchronized entities

+++ src/main/java/com/fedag/CSR/dto/request/BalanceRequest.java
@@ -1,11 +1,13 @@
 package com.fedag.CSR.dto.request;
 
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
 public class BalanceRequest {
-    private Integer id;
+    private BigDecimal id;
     private int coins;
-    private List<Integer> balanceItemsId;
+    private List<BigDecimal> balanceItemsId;
 }
