+++ src/main/java/com/fedag/CSR/service/BalanceService.java
@@ -5,6 +5,8 @@ import com.fedag.CSR.mapper.BalanceMapper;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
+import java.math.BigDecimal;
+
 /**
  * class CommentMapperImpl is implementation of {@link BalanceMapper} interface.
  *
@@ -13,14 +15,14 @@ import org.springframework.data.domain.Pageable;
  */
 
 public interface BalanceService {
-    Balance findById(Integer id);
+    Balance findById(BigDecimal id);
 
     Page<Balance> findAll(Pageable pageable);
 
     Balance create(Balance companyDto);
 
-    Balance update(Integer id, Balance companyDto);
+    Balance update(BigDecimal id, Balance companyDto);
 
-    void deleteById(Integer id);
+    void deleteById(BigDecimal id);
 
 }
